# MQE: Add support for increase function