--- a/pkg/streamingpromql/testdata/ours/native_histograms.test
+++ b/pkg/streamingpromql/testdata/ours/native_histograms.test
@@ -102,12 +102,15 @@ eval range from 0 to 8m step 1m histogram_sum(mixed_metric)
 
 clear
 
-# Test mixed schema rate
+# Test mixed schema rate and increase
 load 1m
 	incr_histogram	{{schema:3 sum:4 count:4 buckets:[1 2 1]}}+{{schema:5 sum:2 count:1 buckets:[1] offset:1}}x10
 
 eval instant at 5m rate(incr_histogram[5m])
 	{} {{schema:3 count:0.016666666666666666 sum:0.03333333333333333 offset:1 buckets:[0.016666666666666666]}}
+
+eval instant at 5m increase(incr_histogram[5m])
+	{} {{schema:3 count:5 sum:10 offset:1 buckets:[5]}}
 
 clear
 
@@ -119,18 +122,24 @@ load 1m
 eval_warn instant at 5m rate(incr_histogram[5m])
   # We expect no results.
 
+eval_warn instant at 5m increase(incr_histogram[5m])
+  # We expect no results.
+
 clear
 
 # Test mixed metrics and range query
 load 1m
 	incr_histogram	1 2 {{schema:3 sum:4 count:4 buckets:[1 2 1]}}+{{schema:5 sum:2 count:1 buckets:[1] offset:1}}x3
 
-# - The first value isn't enough to calculate a rate
-# - The second value is a rate from two floats
-# - The third value is a rate across a float and histogram (so no value returned)
-# - The remaining values contain the rate across two histograms in the vector
+# - The first value isn't enough to calculate a rate/increase
+# - The second value is a rate/increase from two floats
+# - The third value is a rate/increase across a float and histogram (so no value returned)
+# - The remaining values contain the rate/increase across two histograms in the vector
 eval_warn range from 0 to 4m step 1m rate(incr_histogram[1m])
 	{} _ 0.016666666666666666 _ {{schema:3 count:0.016666666666666666 sum:0.03333333333333333 offset:1 buckets:[0.016666666666666666]}} {{schema:3 count:0.016666666666666666 sum:0.03333333333333333 offset:1 buckets:[0.016666666666666666]}}
+
+eval_warn range from 0 to 4m step 1m increase(incr_histogram[1m])
+	{} _ 1 _ {{schema:3 count:1 sum:2 offset:1 buckets:[1]}} {{schema:3 count:1 sum:2 offset:1 buckets:[1]}}
 
 clear
 
@@ -145,14 +154,21 @@ eval range from 0 to 4m step 1m sum by (env) (rate(incr_histogram[5m]))
 	{env="a"}	_ {{count:0.01 sum:0.02 offset:1 buckets:[0.01]}} {{count:0.016666666666666666 sum:0.03333333333333333 offset:1 buckets:[0.016666666666666666]}} {{count:0.023333333333333334 sum:0.04666666666666667 offset:1 buckets:[0.023333333333333334]}} {{count:0.03333333333333333 sum:0.06666666666666667 offset:1 buckets:[0.03333333333333333]}}
 	{env="b"}	_ {{count:0.01 sum:0.02 offset:1 buckets:[0.01]}} {{count:0.016666666666666666 sum:0.03333333333333333 offset:1 buckets:[0.016666666666666666]}} {{count:0.023333333333333334 sum:0.04666666666666667 offset:1 buckets:[0.023333333333333334]}} {{count:0.03333333333333333 sum:0.06666666666666667 offset:1 buckets:[0.03333333333333333]}}
 
+eval range from 0 to 4m step 1m sum by (env) (increase(incr_histogram[5m]))
+	{env="a"}	_ {{count:3 sum:6 offset:1 buckets:[3]}} {{count:5 sum:10 offset:1 buckets:[5]}} {{count:7 sum:14 offset:1 buckets:[7]}} {{count:10 sum:20 offset:1 buckets:[10]}}
+	{env="b"}	_ {{count:3 sum:6 offset:1 buckets:[3]}} {{count:5 sum:10 offset:1 buckets:[5]}} {{count:7 sum:14 offset:1 buckets:[7]}} {{count:10 sum:20 offset:1 buckets:[10]}}
+
 clear
 
-# Test rate() with explicit counter reset information.
+# Test rate/increase with explicit counter reset information.
 load 1m
   metric {{sum:3 count:4 buckets:[1 2 1] counter_reset_hint:reset}} {{sum:63 count:124 buckets:[100 4 20] counter_reset_hint:not_reset}}
 
 eval instant at 1m rate(metric[1m])
   {} {{sum:1 count:2 buckets:[1.65 0.03333333333333333 0.31666666666666665] counter_reset_hint:gauge}}
+
+eval instant at 1m increase(metric[1m])
+  {} {{sum:60 count:120 buckets:[99 2 19] counter_reset_hint:gauge}}
 
 clear