--- a/pkg/streamingpromql/testdata/upstream/native_histograms.test
+++ b/pkg/streamingpromql/testdata/upstream/native_histograms.test
@@ -933,18 +933,15 @@ clear
 load 5m
      reset_in_bucket {{schema:0 count:4 sum:5 buckets:[1 2 1]}} {{schema:0 count:5 sum:6 buckets:[1 1 3]}} {{schema:0 count:6 sum:7 buckets:[1 2 3]}}
 
-# Unsupported by streaming engine.
-# eval instant at 10m increase(reset_in_bucket[15m])
-#      {} {{count:9 sum:10.5 buckets:[1.5 3 4.5]}}
+eval instant at 10m increase(reset_in_bucket[15m])
+     {} {{count:9 sum:10.5 buckets:[1.5 3 4.5]}}
 
 # The following two test the "fast path" where only sum and count is decoded.
-# Unsupported by streaming engine.
-# eval instant at 10m histogram_count(increase(reset_in_bucket[15m]))
-#      {} 9
+eval instant at 10m histogram_count(increase(reset_in_bucket[15m]))
+     {} 9
 
-# Unsupported by streaming engine.
-# eval instant at 10m histogram_sum(increase(reset_in_bucket[15m]))
-#      {} 10.5
+eval instant at 10m histogram_sum(increase(reset_in_bucket[15m]))
+     {} 10.5
 
 clear