@jhesketh 2024-09-26
Will update changelog once https://github.com/grafana/mimir/pull/9419 merges