@jhesketh 2024-09-26
I won't add benchmarks for `increase` since it should be pretty much identical to `rate`. I'll do some benchmarks to check for regressions though.