@jhesketh 2024-09-26
Here are the benchmark results vs main:

<details>

```

goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
cpu: Intel(R) Xeon(R) CPU E5-2686 v4 @ 2.30GHz
                                                                                                                             │   main.txt    │           increase.txt            │
                                                                                                                             │    sec/op     │   sec/op     vs base              │
Query/rate(a_1[1m]),_instant_query/engine=Mimir-8                                                                               645.5µ ±  2%   657.5µ ± 1%  +1.85% (p=0.015 n=6)
Query/rate(a_1[1m]),_instant_query/engine=Prometheus-8                                                                          675.3µ ±  2%   686.5µ ± 1%  +1.66% (p=0.015 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps/engine=Mimir-8                                                                  696.4µ ±  1%   713.0µ ± 0%  +2.38% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                             747.3µ ±  2%   752.9µ ± 2%       ~ (p=0.310 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                                 881.2µ ±  1%   897.2µ ± 1%  +1.81% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                            1.013m ±  1%   1.028m ± 1%  +1.41% (p=0.009 n=6)
Query/rate(a_100[1m]),_instant_query/engine=Mimir-8                                                                             1.309m ± 19%   1.327m ± 0%       ~ (p=0.093 n=6)
Query/rate(a_100[1m]),_instant_query/engine=Prometheus-8                                                                        1.525m ±  1%   1.549m ± 2%  +1.57% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps/engine=Mimir-8                                                                4.108m ±  0%   4.235m ± 0%  +3.07% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                           5.111m ±  1%   5.166m ± 1%  +1.09% (p=0.026 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                               19.80m ±  2%   20.04m ± 1%  +1.24% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                          30.62m ±  1%   30.52m ± 1%       ~ (p=0.180 n=6)
Query/rate(a_2000[1m]),_instant_query/engine=Mimir-8                                                                            10.54m ± 23%   10.35m ± 0%       ~ (p=0.093 n=6)
Query/rate(a_2000[1m]),_instant_query/engine=Prometheus-8                                                                       13.98m ±  1%   13.99m ± 0%       ~ (p=0.485 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps/engine=Mimir-8                                                               65.81m ±  2%   66.90m ± 1%  +1.66% (p=0.026 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                          85.73m ±  1%   86.70m ± 1%  +1.13% (p=0.009 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                              366.6m ±  1%   376.3m ± 1%  +2.63% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                         583.3m ±  0%   585.2m ± 1%       ~ (p=0.180 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                                2.740m ±  1%   2.785m ± 0%  +1.62% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                           3.778m ±  1%   3.778m ± 0%       ~ (p=0.818 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                              187.3m ±  1%   191.9m ± 1%  +2.45% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                         292.7m ±  1%   294.7m ± 1%       ~ (p=0.065 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                              3.750 ±  0%    3.827 ± 0%  +2.05% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                         6.162 ±  0%    6.176 ± 1%       ~ (p=0.132 n=6)
Query/rate(nh_1[1m]),_instant_query/engine=Mimir-8                                                                              676.0µ ±  1%   677.0µ ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_1[1m]),_instant_query/engine=Prometheus-8                                                                         700.2µ ±  1%   708.8µ ± 1%  +1.24% (p=0.004 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/engine=Mimir-8                                                                 1.025m ±  1%   1.025m ± 1%       ~ (p=1.000 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                           1001.4µ ±  1%   999.0µ ± 1%       ~ (p=0.485 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                                1.923m ±  1%   1.951m ± 2%  +1.41% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                           2.309m ±  2%   2.306m ± 2%       ~ (p=0.818 n=6)
Query/rate(nh_100[1m]),_instant_query/engine=Mimir-8                                                                            1.751m ±  3%   1.722m ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_100[1m]),_instant_query/engine=Prometheus-8                                                                       2.007m ±  1%   2.010m ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/engine=Mimir-8                                                               31.23m ±  1%   31.26m ± 0%       ~ (p=0.394 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                          25.03m ±  1%   25.20m ± 1%  +0.68% (p=0.041 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                              120.0m ±  1%   120.5m ± 1%       ~ (p=0.132 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                         153.9m ±  1%   154.5m ± 1%  +0.44% (p=0.041 n=6)
Query/rate(nh_2000[1m]),_instant_query/engine=Mimir-8                                                                           17.00m ±  1%   16.90m ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1m]),_instant_query/engine=Prometheus-8                                                                      21.94m ±  3%   21.81m ± 1%       ~ (p=0.180 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/engine=Mimir-8                                                              609.5m ±  1%   609.6m ± 2%       ~ (p=0.818 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/engine=Prometheus-8                                                         482.0m ±  0%   482.3m ± 1%       ~ (p=0.699 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                              2.700 ±  1%    2.699 ± 0%       ~ (p=1.000 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                                         3.320 ±  1%    3.333 ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                               12.73m ±  2%   12.94m ± 1%  +1.62% (p=0.009 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                          16.21m ±  2%   16.11m ± 2%       ~ (p=0.937 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                              1.351 ±  1%    1.354 ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                         1.676 ±  1%    1.679 ± 0%       ~ (p=0.240 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/engine=Mimir-8                                                             26.40 ±  0%    26.37 ± 0%       ~ (p=0.818 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/engine=Prometheus-8                                                        32.99 ±  1%    33.13 ± 0%       ~ (p=0.065 n=6)
Query/rate(a_1[1d]),_instant_query/engine=Mimir-8                                                                               1.916m ±  3%   1.921m ± 1%       ~ (p=0.485 n=6)
Query/rate(a_1[1d]),_instant_query/engine=Prometheus-8                                                                          2.292m ±  2%   2.296m ± 1%       ~ (p=0.699 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps/engine=Mimir-8                                                                  2.692m ±  0%   2.710m ± 1%  +0.66% (p=0.026 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps/engine=Prometheus-8                                                             3.557m ±  2%   3.512m ± 1%       ~ (p=0.132 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps/engine=Mimir-8                                                                 9.452m ±  1%   9.441m ± 0%       ~ (p=0.394 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps/engine=Prometheus-8                                                            14.49m ±  3%   14.64m ± 2%       ~ (p=0.394 n=6)
Query/rate(a_100[1d]),_instant_query/engine=Mimir-8                                                                             100.2m ±  1%   100.7m ± 1%       ~ (p=0.132 n=6)
Query/rate(a_100[1d]),_instant_query/engine=Prometheus-8                                                                        119.2m ±  1%   119.1m ± 1%       ~ (p=0.699 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps/engine=Mimir-8                                                                178.5m ±  1%   179.0m ± 0%       ~ (p=0.485 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps/engine=Prometheus-8                                                           240.8m ±  1%   239.5m ± 1%       ~ (p=0.180 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps/engine=Mimir-8                                                               850.6m ±  2%   845.5m ± 1%       ~ (p=0.132 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps/engine=Prometheus-8                                                           1.315 ±  5%    1.313 ± 1%       ~ (p=0.818 n=6)
Query/rate(a_2000[1d]),_instant_query/engine=Mimir-8                                                                             1.851 ±  3%    1.862 ± 1%       ~ (p=0.394 n=6)
Query/rate(a_2000[1d]),_instant_query/engine=Prometheus-8                                                                        2.215 ±  1%    2.221 ± 0%       ~ (p=0.132 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps/engine=Mimir-8                                                                3.388 ±  1%    3.404 ± 0%       ~ (p=0.093 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps/engine=Prometheus-8                                                           4.595 ±  3%    4.662 ± 4%       ~ (p=0.093 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps/engine=Mimir-8                                                               16.88 ±  1%    16.79 ± 1%       ~ (p=0.093 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps/engine=Prometheus-8                                                          25.95 ±  3%    25.91 ± 1%       ~ (p=0.394 n=6)
Query/rate(nh_1[1h]),_instant_query/engine=Mimir-8                                                                              1.031m ±  0%   1.042m ± 1%  +1.02% (p=0.041 n=6)
Query/rate(nh_1[1h]),_instant_query/engine=Prometheus-8                                                                         1.083m ±  1%   1.083m ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/engine=Mimir-8                                                                 1.330m ±  1%   1.338m ± 1%       ~ (p=0.065 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/engine=Prometheus-8                                                            1.776m ±  1%   1.784m ± 1%       ~ (p=0.699 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/engine=Mimir-8                                                                4.435m ±  1%   4.444m ± 1%       ~ (p=0.132 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/engine=Prometheus-8                                                           7.780m ±  1%   7.825m ± 1%       ~ (p=0.065 n=6)
Query/rate(nh_100[1h]),_instant_query/engine=Mimir-8                                                                            24.11m ±  1%   24.12m ± 1%       ~ (p=0.394 n=6)
Query/rate(nh_100[1h]),_instant_query/engine=Prometheus-8                                                                       23.05m ±  1%   23.16m ± 1%       ~ (p=0.180 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/engine=Mimir-8                                                               54.21m ±  0%   54.16m ± 0%       ~ (p=0.937 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/engine=Prometheus-8                                                          86.87m ±  1%   86.86m ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/engine=Mimir-8                                                              357.3m ±  1%   361.4m ± 1%  +1.13% (p=0.004 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/engine=Prometheus-8                                                         691.3m ±  2%   696.4m ± 4%       ~ (p=0.485 n=6)
Query/rate(nh_2000[1h]),_instant_query/engine=Mimir-8                                                                           457.6m ±  1%   454.1m ± 1%       ~ (p=0.093 n=6)
Query/rate(nh_2000[1h]),_instant_query/engine=Prometheus-8                                                                      434.3m ±  1%   434.7m ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/engine=Mimir-8                                                               1.090 ±  0%    1.099 ± 2%  +0.82% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/engine=Prometheus-8                                                          1.738 ±  4%    1.736 ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/engine=Mimir-8                                                              7.551 ±  0%    7.569 ± 2%       ~ (p=0.394 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/engine=Prometheus-8                                                         13.88 ±  1%    13.86 ± 1%       ~ (p=1.000 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query/engine=Mimir-8                                                               1.210m ±  1%   1.229m ± 1%  +1.58% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query/engine=Prometheus-8                                                          1.257m ±  1%   1.282m ± 1%  +2.00% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps/engine=Mimir-8                                                  1.313m ±  1%   1.342m ± 2%  +2.22% (p=0.004 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps/engine=Prometheus-8                                             1.439m ±  1%   1.462m ± 1%  +1.57% (p=0.026 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps/engine=Mimir-8                                                 1.661m ±  1%   1.674m ± 1%  +0.79% (p=0.026 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                            2.373m ±  0%   2.387m ± 1%  +0.60% (p=0.009 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query/engine=Mimir-8                                                           2.646m ±  1%   2.676m ± 1%       ~ (p=0.093 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query/engine=Prometheus-8                                                      2.945m ±  1%   2.954m ± 0%       ~ (p=0.394 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps/engine=Mimir-8                                              8.341m ±  1%   8.445m ± 0%  +1.25% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps/engine=Prometheus-8                                         15.40m ±  1%   15.41m ± 1%       ~ (p=1.000 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps/engine=Mimir-8                                             39.48m ±  1%   40.34m ± 0%  +2.18% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                        113.8m ±  2%   112.0m ± 1%       ~ (p=0.093 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query/engine=Mimir-8                                                         23.90m ±  1%   23.91m ± 1%       ~ (p=0.699 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query/engine=Prometheus-8                                                    31.30m ±  1%   31.21m ± 1%       ~ (p=0.394 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps/engine=Mimir-8                                            137.4m ±  1%   138.9m ± 1%  +1.14% (p=0.041 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps/engine=Prometheus-8                                       305.9m ±  1%   307.7m ± 2%       ~ (p=0.093 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps/engine=Mimir-8                                           752.3m ±  1%   771.0m ± 2%  +2.49% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps/engine=Prometheus-8                                       2.586 ±  2%    2.600 ± 1%       ~ (p=0.589 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_instant_query/engine=Mimir-8                                                                  726.0µ ±  1%   727.6µ ± 1%       ~ (p=0.180 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_instant_query/engine=Prometheus-8                                                             766.9µ ±  1%   764.9µ ± 1%       ~ (p=0.485 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_100_steps/engine=Mimir-8                                                     959.0µ ±  0%   974.2µ ± 1%  +1.59% (p=0.004 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_100_steps/engine=Prometheus-8                                                1.058m ±  1%   1.070m ± 1%  +1.11% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1000_steps/engine=Mimir-8                                                    2.040m ±  1%   2.079m ± 1%  +1.89% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1000_steps/engine=Prometheus-8                                               2.818m ±  0%   2.826m ± 0%       ~ (p=0.132 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_instant_query/engine=Mimir-8                                                                4.026m ±  3%   4.019m ± 1%       ~ (p=0.937 n=6)
…